requirements.txt
feedparser>=6.0
pyyaml>=6.0
anthropic>=0.40.0
# openai>=1.0  # uncomment if using OpenAI provider
